- fixed ReadFieldLength to return a long so bigint autoincrement columns can work (bug #58373)

+++ MySql.Data/Provider/Source/Types/MySqlBit.cs
@@ -109,7 +109,7 @@ namespace MySql.Data.Types
 
 		public void SkipValue(MySqlPacket packet)
 		{
-            int len = packet.ReadFieldLength();
+            int len = (int)packet.ReadFieldLength();
             packet.Position += len;
 		}
